chatbot/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
langgraph>=0.1.0
psycopg[binary]>=3.0.0